// src/api/generated/models/companyMember.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { CompanyMemberRole } from './companyMemberRole';
import type { CompanyMemberStatus } from './companyMemberStatus';
import type { CompanyMemberPermissions } from './companyMemberPermissions';
import type { CompanyMemberEffectivePermissionsItem } from './companyMemberEffectivePermissionsItem';
import type { CompanyMemberInvitedAt } from './companyMemberInvitedAt';
import type { CompanyMemberJoinedAt } from './companyMemberJoinedAt';
import type { AuthUser } from './authUser';

export interface CompanyMember {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  role: CompanyMemberRole;
  status: CompanyMemberStatus;
  permissions: CompanyMemberPermissions;
  effectivePermissions: CompanyMemberEffectivePermissionsItem[];
  invitedAt: CompanyMemberInvitedAt;
  joinedAt: CompanyMemberJoinedAt;
  user?: AuthUser;
}